add butterworth and kalman filter

+++ in4073/fixedpoint.h
@@ -8,6 +8,6 @@
 int32_t flo2fix(float value);
 int32_t _16to32(int16_t value);
 int16_t _32to16(int32_t value);
-int32_t multiply(int32_t a, int32_t b);
+int64_t multiply(int32_t a, int32_t b);
 
 #endif
